+++ sandbox/rebuild.ts
@@ -1,15 +1,13 @@
-import fs from 'fs';
-import tar from 'tar';
+import { getAnvilCommand, run as installFoundry } from '@foundry-rs/easy-foundryup';
 import { NameService, mergeNetworkConfig } from '../src/config';
 import { Deployer } from '../src/deployer';
-import { unwrap } from '../src/utils';
-import { Sandbox } from '.';
+import { Sandbox } from './index';
 
 // NOTE: Node module execution is used for rebuilding sandbox package
 if (require.main === module) {
   (async () => {
     try {
-      const sandbox = await Sandbox.create({ extract: false });
+      const sandbox = await Sandbox.create({ rebuild: true, verbose: true });
       await sandbox.start({ noSnapshot: true });
 
       const unsDeployer = await Deployer.create();
@@ -20,21 +18,10 @@ if (require.main === module) {
 
       const unsConfig = await unsDeployer.execute(['full', 'config_polygon_pos_bridge']);
       const ensConfig = await ensDeployer.execute(['ens', 'ens_custody', 'fund_ens_custody']);
-      await sandbox.stop();
+      sandbox.stop();
 
       mergeNetworkConfig(unsConfig, NameService.UNS);
       mergeNetworkConfig(ensConfig, NameService.ENS);
-
-      const { dbPath, snapshotPath } = unwrap(sandbox.options, 'network');
-      await tar.create(
-        {
-          cwd: dbPath,
-          gzip: true,
-          file: snapshotPath,
-          filter: (p) => p.indexOf('_tmp') === -1,
-        },
-        fs.readdirSync(dbPath),
-      );
     } catch (error) {
       console.error(error);
     }