+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## v0.9.14
+
+- Add SeaportProxyBuyer contract
+- Update sandbox RPC server to Foundry Anvil to support cancun hardfork
+- Add Seaport contracts to sandbox
+
 ## v0.9.13
 
 - Added `USDT` MATIC token to resolver list